redesigned project structure

+++ src/cli/commands/clear.py
@@ -3,28 +3,31 @@ from pathlib import Path
 
 from utils import ProjectTree
 
-def _rm_app_dist(dist_dir):
+def _rm_app_dist(root):
     """ """
-    for path in Path(dist_dir).iterdir():
+    tree = ProjectTree(root)
+    for path in tree.app_dist.iterdir():
         if path.name != "responses":
             if path.is_file():
                 path.unlink()
             else:
                 shutil.rmtree(path)
 
-def _rm_saved_models(model_dir):
+def _rm_saved_models(root):
     """ """
-    for path in Path(model_dir).iterdir():
-        shutil.rmtree(path)
+    tree = ProjectTree(root)
+    for path in tree.saved_models.iterdir():
+        if path.is_dir():
+            shutil.rmtree(path)
 
 def clear(front, back, root):
     """ """
     tree = ProjectTree(root)
 
     if front and not back:
-        _rm_app_dist(tree.app_dist)
+        _rm_app_dist(root)
     elif back and not front:
-        _rm_saved_models(tree.models)
+        _rm_saved_models(root)
     else:
-        _rm_app_dist(tree.app_dist)
-        _rm_saved_models(tree.models)
+        _rm_app_dist(root)
+        _rm_saved_models(root)
